refactor: move SchedulePoll trait to cc package
