+++ packages/host/src/ZWaveHost.ts
@@ -10,7 +10,6 @@ import type {
 	SecurityManagers,
 	SendCommandOptions,
 	SendCommandReturnType,
-	ValueID,
 } from "@zwave-js/core";
 import type { ZWaveHostOptions } from "./ZWaveHostOptions.js";
 
@@ -85,22 +84,3 @@ export interface GetCommunicationTimeouts {
 }
 
 export type LogNode = Pick<ControllerLogger, "logNode">;
-
-/** Allows scheduling a value refresh (poll) for a later time */
-export interface SchedulePoll {
-	schedulePoll(
-		nodeId: number,
-		valueId: ValueID,
-		options: NodeSchedulePollOptions,
-	): boolean;
-}
-
-export interface NodeSchedulePollOptions {
-	/** The timeout after which the poll is to be scheduled */
-	timeoutMs?: number;
-	/**
-	 * The expected value that's should be verified with this poll.
-	 * When this value is received in the meantime, the poll will be cancelled.
-	 */
-	expectedValue?: unknown;
-}
